Excise the reconciler's rebuild pass
Deletes the P1-P7 image reconciliation chain, park, submission and
acceptance stamping. An existing server now reconciles to itself;
image drift is unreachable behind the API guard. The re-implementation
builds from the machine-checked protocol spec's decision table.

Also deletes serverRebuildStateRank/advanceRebuildState (carried
forward from task 3: this task's functions were their only remaining
callers) and, since no non-test caller of the client interface's
RebuildServer method remained, the ServerInterface method,
ServerRebuildOptions type, and ComputeClient's concrete implementation
(compute.go) plus their tests (TestRebuildServerSendsOnlyImage and the
stub client's rebuild method) -- left dead code otherwise. Regenerates
the ServerInterface/ComputeInterface mocks accordingly.
