+++ pkg/providers/internal/openstack/mock/interfaces.go
@@ -2117,21 +2117,6 @@ func (mr *MockServerInterfaceMockRecorder) RebootServer(ctx, id, hard any) *gomo
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RebootServer", reflect.TypeOf((*MockServerInterface)(nil).RebootServer), ctx, id, hard)
 }
 
-// RebuildServer mocks base method.
-func (m *MockServerInterface) RebuildServer(ctx context.Context, id string, options openstack.ServerRebuildOptions) (*servers.Server, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "RebuildServer", ctx, id, options)
-	ret0, _ := ret[0].(*servers.Server)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// RebuildServer indicates an expected call of RebuildServer.
-func (mr *MockServerInterfaceMockRecorder) RebuildServer(ctx, id, options any) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RebuildServer", reflect.TypeOf((*MockServerInterface)(nil).RebuildServer), ctx, id, options)
-}
-
 // ShowConsoleOutput mocks base method.
 func (m *MockServerInterface) ShowConsoleOutput(ctx context.Context, id string, length *int) (string, error) {
 	m.ctrl.T.Helper()
@@ -2356,21 +2341,6 @@ func (m *MockComputeInterface) RebootServer(ctx context.Context, id string, hard
 func (mr *MockComputeInterfaceMockRecorder) RebootServer(ctx, id, hard any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RebootServer", reflect.TypeOf((*MockComputeInterface)(nil).RebootServer), ctx, id, hard)
-}
-
-// RebuildServer mocks base method.
-func (m *MockComputeInterface) RebuildServer(ctx context.Context, id string, options openstack.ServerRebuildOptions) (*servers.Server, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "RebuildServer", ctx, id, options)
-	ret0, _ := ret[0].(*servers.Server)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// RebuildServer indicates an expected call of RebuildServer.
-func (mr *MockComputeInterfaceMockRecorder) RebuildServer(ctx, id, options any) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RebuildServer", reflect.TypeOf((*MockComputeInterface)(nil).RebuildServer), ctx, id, options)
 }
 
 // ShowConsoleOutput mocks base method.
