+++ pkg/providers/internal/openstack/provider.go
@@ -2154,9 +2154,7 @@ const (
 )
 
 // healthMessageIndeterminate is the Healthy-condition message used whenever a
-// server's health cannot be determined. markServerRebuildAccepted (reconciler)
-// and convertServerHealthStatus's REBUILD branch (monitor) MUST write the same
-// value, or the two writers churn the message on every poll.
+// server's health cannot be determined.
 const healthMessageIndeterminate = "unable to determine server status"
 
 // convertServerHealthStatus translates from an OpenStack server status into a Kubernetes one.
@@ -2180,7 +2178,7 @@ func convertServerHealthStatus(server *servers.Server) (corev1.ConditionStatus, 
 		// A rebuilding server's real state is a lifecycle one, carried on the
 		// Active condition (ActiveConditionReasonRebuilding); it is not serving
 		// during the reimage, so health is reported as indeterminate rather than
-		// asserting a verdict. Message shared with markServerRebuildAccepted.
+		// asserting a verdict.
 		return corev1.ConditionUnknown, unikornv1core.ConditionReasonUnknown, healthMessageIndeterminate
 	case novaStatusUnknown:
 		return corev1.ConditionUnknown, unikornv1core.ConditionReasonUnknown, "unable to determine server status"
@@ -2567,343 +2565,6 @@ func (p *Provider) reconcileLoadBalancerFloatingIP(ctx context.Context, client F
 	return nil
 }
 
-// serverRebuildStateRank orders the rebuild states for the forward-only
-// advancement rule: Initiated < Rebuilding < Succeeded == Failed. The two
-// terminals share a rank, so neither can supersede the other.
-func serverRebuildStateRank(state unikornv1.ServerRebuildState) int {
-	switch state {
-	case unikornv1.ServerRebuildStateInitiated:
-		return 1
-	case unikornv1.ServerRebuildStateRebuilding:
-		return 2
-	case unikornv1.ServerRebuildStateSucceeded, unikornv1.ServerRebuildStateFailed:
-		return 3
-	default:
-		return 0
-	}
-}
-
-// advanceRebuildState advances the marker to state unless the forward-only rank
-// check makes it a no-op. Every reconciler- and monitor-side advance goes
-// through here, so a late or duplicate observation can never retreat a state.
-func advanceRebuildState(rebuild *unikornv1.ServerRebuildStatus, state unikornv1.ServerRebuildState) {
-	if serverRebuildStateRank(state) <= serverRebuildStateRank(rebuild.State) {
-		return
-	}
-
-	rebuild.State = state
-}
-
-// serverRebuildTaskActive reports whether a rebuild-relevant operation is in
-// flight. task_state is the authoritative signal: non-empty for the whole
-// rebuild window, empty at rest. This is what lets a converged read mean
-// completion: Nova flips the image ref to the target and sets task_state
-// together at accept, so (ref == target ∧ task_state empty) is unobservable
-// while a rebuild is in flight and therefore uniquely means it is done. REBUILD
-// is folded in defensively — Nova projects it only from an active rebuild task,
-// so REBUILD with an empty task_state is unreachable, and if seen is treated as
-// active, never as quiescence.
-func serverRebuildTaskActive(openstackServer *servers.Server) bool {
-	return openstackServer.TaskState != "" || openstackServer.Status == novaStatusRebuild
-}
-
-func openstackServerImageID(server *servers.Server) (regionids.ImageID, bool) {
-	value, ok := server.Image["id"].(string)
-	if !ok || value == "" {
-		return regionids.ImageID{}, false
-	}
-
-	imageID, err := regionids.ParseImageID(value)
-	if err != nil {
-		return regionids.ImageID{}, false
-	}
-
-	return imageID, true
-}
-
-// markServerRebuildAccepted stamps the post-acceptance in-flight view: Active
-// Rebuilding and health Unknown, byte-identical to what the monitor derives for
-// a Nova REBUILD (setServerActive + convertServerHealthStatus), including the
-// health message (healthMessageIndeterminate), so the reconciler and monitor
-// writes agree rather than churning the condition every poll. Never call before
-// Nova has accepted: pre-acceptance waits are silent yields and the monitor owns
-// observed state.
-func markServerRebuildAccepted(server *unikornv1.Server) {
-	server.SetActiveCondition(unikornv1.ActiveConditionReasonRebuilding)
-	server.SetHealthCondition(corev1.ConditionUnknown, unikornv1core.ConditionReasonUnknown, healthMessageIndeterminate)
-}
-
-// reconcileServerRebuildPark runs the two-phase, write-ahead park (step P4 of
-// reconcileServerImage's pass order), for a marker whose target is the desired
-// image. It returns handled=true when it owns the pass.
-//
-//   - P4a: a marker read back durably Failed issues the terminal
-//     UserActionRequired park (core turns this into Available=Errored).
-//   - P4b: an ERROR with acceptance evidence — durably >= Rebuilding, or the ref
-//     flipped to the target — stamps Failed and yields to confirm by read-back.
-//   - P4c: supersession — an accepted rebuild whose readable ref has moved off
-//     the target while the task has quiesced can no longer converge; stamp
-//     Failed and yield.
-//
-// P4b's stamp is by direct assignment, the one exception to the forward-only
-// rank rule: it must overwrite a Succeeded stamped moments before the ERROR
-// arrived, or the retained Succeeded level would re-fire the settlement wake
-// forever on the parked server. P4b/P4c precede the unreadable-ref yield (P5)
-// so an ERROR with acceptance evidence parks rather than yielding forever.
-func reconcileServerRebuildPark(ctx context.Context, server *unikornv1.Server, openstackServer *servers.Server, currentImageID regionids.ImageID, refReadable, taskActive bool, desiredImageID regionids.ImageID) (bool, error) {
-	rebuild := server.Status.Rebuild
-	accepted := serverRebuildStateRank(rebuild.State) >= serverRebuildStateRank(unikornv1.ServerRebuildStateRebuilding)
-	converged := refReadable && currentImageID == desiredImageID
-
-	// P4a: Failed read back durable — issue the confirmed terminal park.
-	if rebuild.State == unikornv1.ServerRebuildStateFailed {
-		setServerHealthStatus(server, openstackServer)
-		setServerActive(ctx, server, openstackServer, nil)
-
-		return true, provisioners.UserActionRequired("server_rebuild_failed", "server entered an error state after a rebuild was issued; select another image or replace the server")
-	}
-
-	// P4b: ERROR with acceptance evidence — stamp Failed (direct) and yield.
-	if openstackServer.Status == novaStatusError && (accepted || converged) {
-		setServerHealthStatus(server, openstackServer)
-		setServerActive(ctx, server, openstackServer, nil)
-
-		rebuild.State = unikornv1.ServerRebuildStateFailed
-
-		log.FromContext(ctx).Info("recording the server rebuild failure",
-			"server", server.Name, "novaServerID", openstackServer.ID)
-
-		return true, provisioners.ErrYield
-	}
-
-	// P4c: supersession — accepted, readable ref off target, task quiesced.
-	if accepted && refReadable && !converged && !taskActive {
-		setServerHealthStatus(server, openstackServer)
-		setServerActive(ctx, server, openstackServer, nil)
-
-		rebuild.State = unikornv1.ServerRebuildStateFailed
-
-		log.FromContext(ctx).Info("recording the superseded server rebuild failure",
-			"server", server.Name, "novaServerID", openstackServer.ID)
-
-		return true, provisioners.ErrYield
-	}
-
-	return false, nil
-}
-
-// submitServerRebuild issues the Nova rebuild and, on acceptance, advances the
-// marker to Rebuilding (step P7e of reconcileServerImagePending). A 409 is
-// pre-acceptance: it yields for a short retry with the state left at Initiated,
-// silently — the server is still running untouched, so no Phase/Healthy write
-// fights the monitor's observed state.
-//
-// On acceptance it writes the fixed accepted stamp (markServerRebuildAccepted)
-// and never derives Phase/Healthy from the rebuild response body: a 202 body
-// can still read ACTIVE for the pre-destruction server and would falsely stamp
-// the just-accepted destructive rebuild as running.
-func submitServerRebuild(ctx context.Context, client ServerInterface, server *unikornv1.Server, openstackServer *servers.Server, options ServerRebuildOptions) (*servers.Server, error) {
-	rebuilt, err := client.RebuildServer(ctx, openstackServer.ID, options)
-	if err != nil {
-		if gophercloud.ResponseCodeIs(err, http.StatusConflict) {
-			log.FromContext(ctx).Info("server rebuild refused pending another operation, waiting for a rebuildable state",
-				"server", server.Name, "novaServerID", openstackServer.ID)
-
-			return openstackServer, provisioners.ErrYield
-		}
-
-		return nil, err
-	}
-
-	advanceRebuildState(server.Status.Rebuild, unikornv1.ServerRebuildStateRebuilding)
-	markServerRebuildAccepted(server)
-
-	if rebuilt != nil {
-		return rebuilt, nil
-	}
-
-	return openstackServer, nil
-}
-
-// reconcileServerImage converges the server onto its desired image, deciding
-// only from the fresh openstackServer plus the durable marker read back at pass
-// start. It implements the normative pass order (P0 deletion is handled before
-// this by the deprovision path):
-//
-//	P1  Spec.Image == nil        → clear any marker; done.
-//	P3  re-arm (marker target ≠ desired) → replace marker {desired, Initiated}; yield.
-//	P4  two-phase park (reconcileServerRebuildPark).
-//	P5  unreadable ref           → silent yield (never clean-complete over an
-//	                               unverifiable image).
-//	P6  ref == desired           → reconcileServerImageConverged.
-//	P7  ref != desired           → reconcileServerImagePending.
-//
-// Arming and submitting are separate passes with a yield between them as the
-// durable commit point, so intent is persisted before Nova is asked to destroy
-// the root disk; settlement writes are confirmed by read-back, never assumed.
-func reconcileServerImage(ctx context.Context, client ServerInterface, server *unikornv1.Server, openstackServer *servers.Server) (*servers.Server, error) {
-	// P1: no desired image — retire any marker.
-	if server.Spec.Image == nil {
-		server.Status.Rebuild = nil
-
-		return openstackServer, nil
-	}
-
-	desiredImageID := server.Spec.Image.ID
-
-	// P3: re-arm. A marker for a different target — including a parked Failed,
-	// the designed recovery — is replaced with a fresh write-ahead Initiated
-	// marker. The yield is silent: nothing has been asked of Nova.
-	if rebuild := server.Status.Rebuild; rebuild != nil && rebuild.TargetImageID != desiredImageID {
-		server.Status.Rebuild = &unikornv1.ServerRebuildStatus{
-			TargetImageID: desiredImageID,
-			State:         unikornv1.ServerRebuildStateInitiated,
-		}
-
-		log.FromContext(ctx).Info("recording the server rebuild intent",
-			"server", server.Name, "novaServerID", openstackServer.ID)
-
-		return openstackServer, provisioners.ErrYield
-	}
-
-	currentImageID, refReadable := openstackServerImageID(openstackServer)
-	taskActive := serverRebuildTaskActive(openstackServer)
-
-	// P4: two-phase park. Any marker here targets the desired image (P3 handled
-	// a differing target). It precedes P5 so an attributable ERROR with an
-	// unreadable ref parks rather than yielding forever.
-	if server.Status.Rebuild != nil {
-		if handled, err := reconcileServerRebuildPark(ctx, server, openstackServer, currentImageID, refReadable, taskActive, desiredImageID); handled {
-			return openstackServer, err
-		}
-	}
-
-	// P5: every server this provider creates is image-booted, so an unparseable
-	// image ref is abnormal. Yield visibly rather than report success over an
-	// unverifiable image change.
-	if !refReadable {
-		log.FromContext(ctx).Info("cannot read the server image from the provider, image convergence cannot be checked",
-			"server", server.Name, "novaServerID", openstackServer.ID)
-
-		return openstackServer, provisioners.ErrYield
-	}
-
-	// P6: ref converged onto the desired image.
-	if currentImageID == desiredImageID {
-		return reconcileServerImageConverged(ctx, server, openstackServer, taskActive)
-	}
-
-	// P7: ref still pending.
-	return reconcileServerImagePending(ctx, client, server, openstackServer, desiredImageID, taskActive)
-}
-
-// reconcileServerImageConverged runs the ref-converged pass (P6): the fresh
-// read shows the desired image. Any marker here is not Failed (P4a parked it)
-// nor an attributable ERROR (P4b stamped it), so ERROR with a marker is
-// unreachable.
-func reconcileServerImageConverged(ctx context.Context, server *unikornv1.Server, openstackServer *servers.Server, taskActive bool) (*servers.Server, error) {
-	rebuild := server.Status.Rebuild
-
-	// P6c: ERROR with no marker is unrelated — the monitor owns health. (With a
-	// marker this is unreachable, see above.)
-	if openstackServer.Status == novaStatusError {
-		return openstackServer, nil
-	}
-
-	// P6a/P6b with no marker: converged and nothing pending — done.
-	if rebuild == nil {
-		return openstackServer, nil
-	}
-
-	// P6a: quiescent convergence is the settlement — clear the marker and yield
-	// so the requeued pass confirms the clear by read-back.
-	if !taskActive {
-		server.Status.Rebuild = nil
-
-		log.FromContext(ctx).Info("clearing the settled server rebuild marker",
-			"server", server.Name, "novaServerID", openstackServer.ID)
-
-		return openstackServer, provisioners.ErrYield
-	}
-
-	// P6b: target ref, active task. Record the observed acceptance (advance
-	// Rebuilding), re-assert the accepted stamp, and complete — a future monitor
-	// advance to a terminal still exists to wake the settlement pass.
-	if serverRebuildStateRank(rebuild.State) < serverRebuildStateRank(unikornv1.ServerRebuildStateSucceeded) {
-		advanceRebuildState(rebuild, unikornv1.ServerRebuildStateRebuilding)
-		markServerRebuildAccepted(server)
-
-		return openstackServer, nil
-	}
-
-	// A Succeeded marker cannot rank-advance again, so no monitor delta is
-	// guaranteed: yield to carry liveness rather than clean-complete.
-	return openstackServer, provisioners.ErrYield
-}
-
-// reconcileServerImagePending runs the ref-pending pass (P7): the fresh read
-// still shows an image other than the desired one. P4c already handled
-// supersession, so an accepted marker reaching here always has an active task.
-func reconcileServerImagePending(ctx context.Context, client ServerInterface, server *unikornv1.Server, openstackServer *servers.Server, desiredImageID regionids.ImageID, taskActive bool) (*servers.Server, error) {
-	// P7a: rebuild only a server Nova reports as booted. Before first boot the
-	// image is a create parameter and the server is in create-retry's domain;
-	// yield silently to stay visibly provisioning.
-	if openstackServer.LaunchedAt.IsZero() {
-		log.FromContext(ctx).Info("image change deferred until first launch",
-			"server", server.Name, "novaServerID", openstackServer.ID)
-
-		return openstackServer, provisioners.ErrYield
-	}
-
-	rebuild := server.Status.Rebuild
-
-	// P7b: no marker — arm write-ahead intent and yield. The Initiated marker
-	// read back later is the only submission authorization.
-	if rebuild == nil {
-		server.Status.Rebuild = &unikornv1.ServerRebuildStatus{
-			TargetImageID: desiredImageID,
-			State:         unikornv1.ServerRebuildStateInitiated,
-		}
-
-		log.FromContext(ctx).Info("recording the server rebuild intent",
-			"server", server.Name, "novaServerID", openstackServer.ID)
-
-		return openstackServer, provisioners.ErrYield
-	}
-
-	switch rebuild.State {
-	case unikornv1.ServerRebuildStateRebuilding:
-		// P7c: accepted rebuild still in flight — re-assert the accepted stamp
-		// and complete; quiescence lands in P4c, a guaranteed monitor delta.
-		markServerRebuildAccepted(server)
-
-		return openstackServer, nil
-	case unikornv1.ServerRebuildStateSucceeded:
-		// P7c′: a terminal marker guarantees no monitor delta; yield to carry
-		// liveness to P4c.
-		return openstackServer, provisioners.ErrYield
-	case unikornv1.ServerRebuildStateInitiated:
-		// P7d: a task in flight is a foreign/blocking op — yield rather than
-		// clean-complete, or the Initiated marker would have no wake channel.
-		if taskActive {
-			return openstackServer, provisioners.ErrYield
-		}
-
-		// P7e: the only destructive row. A durable Initiated marker with a
-		// quiescent server authorizes exactly one submission.
-		return submitServerRebuild(ctx, client, server, openstackServer, ServerRebuildOptions{
-			ImageID: desiredImageID,
-		})
-	case unikornv1.ServerRebuildStateFailed:
-		// Unreachable: P4a parks a Failed marker before P7. Yield defensively.
-		return openstackServer, provisioners.ErrYield
-	}
-
-	// A state this version does not recognize (version skew during a rolling
-	// upgrade) must never clean-complete as provisioned.
-	return openstackServer, provisioners.ErrYield
-}
-
 func (p *Provider) reconcileServer(ctx context.Context, client ServerInterface, server *unikornv1.Server, port *ports.Port, keyName string, preflight serverCreatePreflight) (*servers.Server, error) {
 	log := log.FromContext(ctx)
 
@@ -2911,7 +2572,10 @@ func (p *Provider) reconcileServer(ctx context.Context, client ServerInterface, 
 	if err == nil {
 		log.V(1).Info("server already exists")
 
-		return reconcileServerImage(ctx, client, server, openstackServer)
+		// Image drift is not acted on: the in-place rebuild is excised and the
+		// API rejects image changes until the re-implementation lands (spec
+		// 2026-07-31-status-observed-implementation-design.md, stage 3).
+		return openstackServer, nil
 	}
 
 	networks := []servers.Network{
